refactor: fix clippy warnings and improve code quality
- Remove unused struct declarations from E2E tests
- Fix module naming conflicts in test files (remove inner mod wrapper)
- Fix benchmark cfg feature name (benchmark -> server)
- Use to_vec() instead of iter().cloned().collect() for clarity
- Auto-fix other minor clippy warnings

All 348 tests pass after these changes.

+++ src/storage/performance.rs
@@ -285,7 +285,7 @@ mod tests {
     #[test]
     fn test_pool_stats_calculation() {
         // Test utilization calculation
-        let util = ((10 - 2) as f64 / 10 as f64) * 100.0;
+        let util = ((10 - 2) as f64 / 10_f64) * 100.0;
         assert_eq!(util, 80.0);
     }
